+++ HISTORY.rst
@@ -6,7 +6,7 @@ History
 2.2.0 (2015-02-26)
 ~~~~~~~~~~~~~~~~~~
 
-- Adding lists of local models as we argument for the local ensemble
+- Adding lists of local models as argument for the local ensemble
   constructor.
 
 2.1.0 (2015-02-22)